+++ src/routes/GroupExplorer/GroupSearch/GroupSearch.test.js
@@ -1,106 +1,67 @@
 import React from 'react'
 import { graphql } from 'msw'
-import { setupServer } from 'msw/node'
-import { history } from 'router'
-import { generateStore, render, AllTheProviders, fireEvent } from 'util/reactTestingLibraryExtended'
-import GroupSearch from './GroupSearch'
+import mockGraphqlServer from 'util/testing/mockGraphqlServer'
 import orm from 'store/models'
+import { render, screen, AllTheProviders } from 'util/testing/reactTestingLibraryExtended'
+import userEvent from '@testing-library/user-event'
+import GroupSearch from './GroupSearch'
 
-jest.mock('components/ScrollListener', () => () => <div />) // was throwing errors with this.element().removeEventListener('blabadlbakdbfl')
+// was throwing errors with this.element().removeEventListener('blabadlbakdbfl')
+jest.mock('components/ScrollListener', () => () => <div />)
 
-const server = setupServer(
-  graphql.operation((req, res, ctx) => {
-    const { search } = req.body.variables
-    return res(ctx.data({
-      groups: {
-        items: search !== 'different group'
-          ? [
-            {
-              accessibility: [3],
-              memberCount: 12,
-              description: 'Words do not belong in this here town',
-              location: 'Baltimore',
-              locationObject: {
-                city: 'Baltimore',
-                country: 'USA',
-                fullText: 'Baltimore, USA',
-                locality: '',
-                neighborhood: '',
-                region: 'East Coast'
-              },
-              id: '345',
-              avatarUrl: 'wee.com',
-              bannerUrl: 'wee.com',
-              name: 'Test Group Title',
-              slug: 'test-group-title',
-              groupTopics: [],
-              members: []
-            }
-          ]
-          : [
-            {
-              accessibility: [3],
-              memberCount: 16,
-              description: 'Completely different group don\'t you think',
-              location: 'Somewhere else',
-              locationObject: {
-                city: 'Austin',
-                country: 'USA',
-                fullText: 'Austin, USA',
-                locality: '',
-                neighborhood: '',
-                region: 'East Coast'
-              },
-              id: '345',
-              avatarUrl: 'wee.com',
-              bannerUrl: 'wee.com',
-              name: 'Search input results',
-              slug: 'test-group-title',
-              groupTopics: [],
-              members: []
-            }
-          ],
-        hasMore: false,
-        total: 0
-      }
-    }))
-  })
-)
+function testProviders () {
+  const ormSession = orm.mutableSession(orm.getEmptyState())
+  ormSession.Me.create({ id: '1' })
+  const reduxState = { orm: ormSession.state }
 
-let providersWithStore = null
-
-// Enable API mocking before tests.
-beforeAll(() => server.listen())
-
-beforeEach(() => {
-  // setup store
-  const session = orm.mutableSession(orm.getEmptyState())
-  session.Me.create({ id: '1' })
-  const initialState = { orm: session.state }
-  const store = generateStore(history, initialState)
-  providersWithStore = AllTheProviders(store)
-})
+  return AllTheProviders(reduxState)
+}
 
 afterEach(() => {
-  server.resetHandlers()
-  providersWithStore = null
+  mockGraphqlServer.resetHandlers()
 })
 
-// Disable API mocking after the tests are done.
-afterAll(() => server.close())
-
 test('GroupSearch integration test', async () => {
-  const { findByText, getByRole } = render(
+  mockGraphqlServer.resetHandlers(
+    graphql.operation((req, res, ctx) => {
+      const { search } = req.body.variables
+      return res(ctx.data({
+        groups: {
+          items: search !== 'different group'
+            ? [
+              {
+                id: '345',
+                name: 'Test Group Title',
+                slug: 'test-group-title',
+                accessibility: [3],
+                description: 'Words do not belong in this here town'
+              }
+            ] : [
+              {
+                id: '345',
+                name: 'Search input results',
+                slug: 'test-group-title',
+                accessibility: [3],
+                description: 'Completely different group don\'t you think'
+              }
+            ],
+          hasMore: false,
+          total: 0
+        }
+      }))
+    })
+  )
+  const user = userEvent.setup()
+
+  render(
     <GroupSearch />,
-    null,
-    providersWithStore
+    { wrapper: testProviders() }
   )
 
-  expect(await findByText('Test Group Title')).toBeInTheDocument()
+  expect(await screen.findByText('Test Group Title')).toBeInTheDocument()
+  expect(screen.queryByText('Search input results')).not.toBeInTheDocument()
 
-  fireEvent.change(getByRole('textbox'), {
-    target: { value: 'different group' }
-  })
+  await user.type(screen.getByRole('textbox'), 'different group')
 
-  expect(await findByText('Search input results')).toBeInTheDocument()
+  expect(await screen.findByText('Search input results')).toBeInTheDocument()
 })