+++ src/routes/MapExplorer/MapExplorer.js
@@ -22,7 +22,6 @@ import LayoutFlagsContext from 'contexts/LayoutFlagsContext'
 import getQuerystringParam from 'store/selectors/getQuerystringParam'
 import { locationObjectToViewport } from 'util/geo'
 import { isMobileDevice } from 'util/mobile'
-import { getQueryParamsObjectFromString } from 'util/navigation'
 import { generateViewParams } from 'util/savedSearch'
 
 import MapDrawer from './MapDrawer'
@@ -104,8 +103,9 @@ export class UnwrappedMapExplorer extends React.Component {
 
     // Relinquishes route handling within the Map entirely to Mobile App
     // e.g. react router / history push
-    const { mobileSettingsLayout } = this.context
-    if (mobileSettingsLayout) {
+    const { hyloAppLayout } = this.context
+
+    if (hyloAppLayout) {
       this.props.history.block(tx => {
         const path = tx.pathname
         // when in embedded view of map allow web navigation within map
@@ -371,7 +371,7 @@ export class UnwrappedMapExplorer extends React.Component {
     setTimeout(() => {
       if (this.state.creatingPost) {
         this.setState({ coordinates: { lng: e.lngLat[0], lat: e.lngLat[1] } })
-        const currentParams = getQueryParamsObjectFromString(this.props.location.search)
+        const currentParams = Object.fromEntries(new URLSearchParams(this.props.location.search))
         this.props.showCreateModal({ ...currentParams, ...this.state.coordinates })
       }
     }, this.state.isAddingItemToMap ? 0 : oneSecondInMs)
@@ -515,8 +515,8 @@ export class UnwrappedMapExplorer extends React.Component {
       viewport
     } = this.state
 
-    const { mobileSettingsLayout } = this.context
-    const withoutNav = mobileSettingsLayout
+    const { hyloAppLayout, hideNavLayout } = this.context
+    const withoutNav = hyloAppLayout || hideNavLayout
 
     const locationParams = this.props['location'] !== undefined ? getQuerystringParam(['zoom', 'center', 'lat', 'lng'], null, this.props) : null
 
@@ -567,6 +567,21 @@ export class UnwrappedMapExplorer extends React.Component {
         <button styleName={cx('toggleFeatureFiltersButton', { open: showFeatureFilters, withoutNav })} onClick={this.toggleFeatureFilters}>
         Features: <strong>{featureTypes.filter(t => filters.featureTypes[t]).length}/{featureTypes.length}</strong>
         </button>
+
+        {currentUser && !hyloAppLayout && (
+          <button
+            data-for='addItemToMapTooltip'
+            data-tip='Add item to map'
+            styleName={cx('addItemToMapButton drawerAdjacentButton', { active: isAddingItemToMap, drawerOpen: !hideDrawer })}
+            onClick={this.handleAddItemToMap}
+          >
+            <Icon name='Plus' styleName={cx({ openDrawer: !hideDrawer, closeDrawer: hideDrawer })} />
+          </button>
+        )}
+        <Tooltip
+          delay={550}
+          id='addItemToMapTooltip' />
+
         {currentUser && <>
           <Icon
             name='Heart'