+++ src/routes/NonAuthLayout/PasswordReset/PasswordReset.connector.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux'
-import { resetPassword } from './PasswordReset.store.js'
-
-export const mapDispatchToProps = {
-  resetPassword
-}
-
-export default connect(null, mapDispatchToProps)