+++ src/routes/NonAuthLayoutRouter/Signup/Signup.store.test.js
@@ -1,0 +1,7 @@
+import { register } from './Signup.store'
+
+describe('register', () => {
+  it('should match latest snapshot', () => {
+    expect(register('name', 'test@hylo.com', 'testPassword')).toMatchSnapshot()
+  })
+})