+++ src/store/actions/changeQuerystringParam.js
@@ -2,6 +2,7 @@ import qs from 'querystring'
 import { addQuerystringToPath } from 'util/navigation'
 import { push, replace } from 'connected-react-router'
 
+// Could have a home in `util/navigation`, or make `util/navigation` a directory and put this in there?
 export default function changeQuerystringParam (props, key, value, defaultValue, useReplace) {
   const querystringParams = qs.parse(props.location.search.substring(1))
   const newQuerystringParams = {