+++ src/store/actions/logout.js
@@ -1,0 +1,10 @@
+import { LOGOUT } from 'store/constants'
+
+export default function logout () {
+  return {
+    type: LOGOUT,
+    payload: {
+      api: { path: '/noo/session', method: 'DELETE' }
+    }
+  }
+}