+++ src/store/index.js
@@ -1,12 +1,9 @@
 import { createStore } from 'redux'
 import createMiddleware from './middleware'
-import orm from './models'
 import rootReducer, { combinedReducers } from './reducers'
 
 export function getEmptyState () {
-  return combinedReducers({
-    orm: orm.getEmptyState()
-  }, { type: '' })
+  return combinedReducers({}, { type: '' })
 }
 
 export default function (history, req) {