+++ CHANGELOG.md
@@ -1,13 +1,23 @@
-# Changelog
+# Changelog 
 All notable changes to Hylo Evo (the Hylo front-end) will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [3.6.0] - 2022-04-28
+
+### Changed
+- Major refactor of authorized vs non-authorized routing, see `RootRouter`, `AuthLayoutRouter`, and `NonAuthLayoutRouter`
+- React Testing Library (RTL) setup simplified and notated
 
 ### Added
+- Changes to authorization and authentication flow to accomodate related API changes
+- Adds new RTL-based tests for new and existing components
 - You can now click on the map to create posts at that location. Either click and hold for a couple seconds, or click the (+) button and then click on the map to add a post at that lat,lng on the map.
+- Discussion posts now can have a location and appear on the map
+
+### Fixed
+- Date/time selector in posts always visible, never disappears below the screen
 
 ## [3.5.2] - 2022-03-28